+++ src/main/java/janggi/piece/Sa.java
@@ -1,9 +1,7 @@
 package janggi.piece;
 
 import janggi.Team.Team;
-import janggi.position.Path;
 import janggi.position.Position;
-import java.util.ArrayList;
 import java.util.List;
 
 public class Sa extends Piece {
@@ -13,22 +11,8 @@ public class Sa extends Piece {
     }
 
     @Override
-    public Path makePath(final Position currentPosition, final Position arrivalPosition) {
-        int differenceForY = arrivalPosition.calculateDifferenceForY(currentPosition);
-        int differenceForX = arrivalPosition.calculateDifferenceForX(currentPosition);
-
-        validateMove(differenceForY, differenceForX);
-
-        final List<Position> positions = new ArrayList<>();
-        int currentY = currentPosition.getY();
-        int currentX = currentPosition.getX();
-        currentY = moveY(arrivalPosition, differenceForY, currentY, positions, currentX);
-        moveX(arrivalPosition, differenceForX, currentX, positions, currentY);
-        return new Path(positions);
-    }
-
-    private int moveY(Position arrivalPosition, int differenceForY, int currentY, List<Position> positions,
-                      int currentX) {
+    public int calculatePathY(Position arrivalPosition, int differenceForY, int differenceForX,
+                              List<Position> positions, int currentY, int currentX) {
         int differenceUnitY = calculateUnit(differenceForY);
         while (currentY != arrivalPosition.getY()) {
             currentY += differenceUnitY;
@@ -37,8 +21,9 @@ public class Sa extends Piece {
         return currentY;
     }
 
-    private int moveX(Position arrivalPosition, int differenceForX, int currentX, List<Position> positions,
-                      int currentY) {
+    @Override
+    public int calculatePathX(Position arrivalPosition, int differenceForY, int differenceForX,
+                              List<Position> positions, int currentY, int currentX) {
         int differenceUnitX = calculateUnit(differenceForX);
         while (currentX != arrivalPosition.getX()) {
             currentX += differenceUnitX;
@@ -47,7 +32,8 @@ public class Sa extends Piece {
         return currentX;
     }
 
-    private void validateMove(int differenceForY, int differenceForX) {
+    @Override
+    public void validateMove(int differenceForY, int differenceForX) {
         if (Math.abs(differenceForY) + Math.abs(differenceForX) != 1) {
             throw new IllegalArgumentException("[ERROR] 사는 한 방향으로 한 칸만 이동할 수 있습니다.");
         }
