+++ src/main/java/janggi/piece/Sang.java
@@ -1,9 +1,7 @@
 package janggi.piece;
 
 import janggi.Team.Team;
-import janggi.position.Path;
 import janggi.position.Position;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
@@ -18,23 +16,8 @@ public class Sang extends Piece {
     }
 
     @Override
-    public Path makePath(final Position currentPosition, final Position arrivalPosition) {
-        int differenceForY = arrivalPosition.calculateDifferenceForY(currentPosition);
-        int differenceForX = arrivalPosition.calculateDifferenceForX(currentPosition);
-
-        validateMove(differenceForY, differenceForX);
-
-        final List<Position> positions = new ArrayList<>();
-        int currentY = currentPosition.getY();
-        int currentX = currentPosition.getX();
-
-        currentY = moveY(arrivalPosition, differenceForY, differenceForX, currentY, positions, currentX);
-        moveX(arrivalPosition, differenceForY, differenceForX, currentY, positions, currentX);
-        return new Path(positions);
-    }
-
-    private int moveY(final Position arrivalPosition, final int differenceForY, final int differenceForX,
-                      int currentY, final List<Position> positions, int currentX) {
+    int calculatePathY(Position arrivalPosition, int differenceForY, int differenceForX,
+                       List<Position> positions, int currentY, int currentX) {
         if (isNotStartDirection(differenceForY)) {
             return currentY;
         }
@@ -51,8 +34,9 @@ public class Sang extends Piece {
         return currentY;
     }
 
-    private int moveX(final Position arrivalPosition, final int differenceForY, final int differenceForX,
-                      int currentY, final List<Position> positions, int currentX) {
+    @Override
+    int calculatePathX(Position arrivalPosition, int differenceForY, int differenceForX,
+                       List<Position> positions, int currentY, int currentX) {
         if (isNotStartDirection(differenceForX)) {
             return currentX;
         }
@@ -69,7 +53,8 @@ public class Sang extends Piece {
         return currentX;
     }
 
-    private void validateMove(final int differenceForY, final int differenceForX) {
+    @Override
+    public void validateMove(final int differenceForY, final int differenceForX) {
         if (canNotMove(differenceForY, differenceForX)) {
             throw new IllegalArgumentException("[ERROR] 상은 직선 1칸 이동 후 대각선 2칸으로만 이동할 수 있습니다.");
         }
